Refactor and unify binary operation tests
- Replace test_binary_ops.cpp and test_elementwise_add.cpp with
test_binary_operations.cpp covering all binary ops (add, sub, mul, div)
- Update CMakeLists.txt to use new test file and target - Improve and
expand test coverage for edge cases and data types - Handle Inf/0 = NaN
in scalar division for IEEE 754 compliance
