+++ tests/kernels/CMakeLists.txt
@@ -12,28 +12,28 @@ target_compile_options(test_cpu_kernel PRIVATE -fno-fast-math)
 gtest_discover_tests(test_cpu_kernel)
 
 # Build Binary Operations tests
-add_executable(test_binary_ops test_binary_ops.cpp)
-target_include_directories(test_binary_ops PRIVATE
+add_executable(test_binary_operations test_binary_operations.cpp)
+target_include_directories(test_binary_operations PRIVATE
                             ${pybind11_SOURCE_DIR}/include
                             ${Python3_INCLUDE_DIRS}
                             ${XSIMD_INCLUDE_DIR})
-target_link_libraries(test_binary_ops gtest_main
+target_link_libraries(test_binary_operations gtest_main
                     cpu_kernel_lib ${Python3_LIBRARIES})
 # Ensure IEEE 754 compliance in tests
-target_compile_options(test_binary_ops PRIVATE -fno-fast-math)
-gtest_discover_tests(test_binary_ops)
+target_compile_options(test_binary_operations PRIVATE -fno-fast-math)
+gtest_discover_tests(test_binary_operations)
 
 # Build Elementwise Add tests
-add_executable(test_elementwise_add test_elementwise_add.cpp)
-target_include_directories(test_elementwise_add PRIVATE
-                            ${pybind11_SOURCE_DIR}/include
-                            ${Python3_INCLUDE_DIRS}
-                            ${XSIMD_INCLUDE_DIR})
-target_link_libraries(test_elementwise_add gtest_main
-                    cpu_kernel_lib ${Python3_LIBRARIES})
-# Ensure IEEE 754 compliance in tests
-target_compile_options(test_elementwise_add PRIVATE -fno-fast-math)
-gtest_discover_tests(test_elementwise_add)
+# add_executable(test_elementwise_add test_elementwise_add.cpp)
+# target_include_directories(test_elementwise_add PRIVATE
+#                             ${pybind11_SOURCE_DIR}/include
+#                             ${Python3_INCLUDE_DIRS}
+#                             ${XSIMD_INCLUDE_DIR})
+# target_link_libraries(test_elementwise_add gtest_main
+#                     cpu_kernel_lib ${Python3_LIBRARIES})
+# # Ensure IEEE 754 compliance in tests
+# target_compile_options(test_elementwise_add PRIVATE -fno-fast-math)
+# gtest_discover_tests(test_elementwise_add)
 
 # Build VecBuffer tests
 add_executable(test_vecbuffer test_vecbuffer.cpp)
